# Faster migration to add the should_track_playtime column 