@tdilauro 2024-01-05
Good catch there! Going from > 300s to < 1s seems like a winner. 🚀